Move helper.quality to helpers
