+++ pytest.ini
@@ -27,8 +27,8 @@ flake8-ignore =
     medusa/helper/common.py D100 D200 D202 D205 D400
     medusa/helper/encoding.py D100 D202 D400 D401
     medusa/helper/exceptions.py D100 D200 D202 D205 D400
-    medusa/helper/quality.py D100 D202 D205 D400
     medusa/helpers/__init__.py D103 N806
+    medusa/helpers/quality.py D100 D202 D205 D400
     medusa/history.py D100 D400 N802 N803 N806
     medusa/image_cache.py D100 D101 D102 D105
     medusa/imdb_popular.py D100
